requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
requests==2.31.0
gunicorn==21.2.0
Werkzeug==3.0.1  # Required by Flask
click==8.1.7     # Required by Flask
itsdangerous==2.1.2  # Required by Flask
Jinja2==3.1.2    # Required by Flask
MarkupSafe==2.1.3  # Required by Jinja2
google-search-results==2.4.2  # SerpApi for web search